Fix wrong cast in SDL2 sample time, see #71.

+++ Samples/basic/sdl2/src/SystemInterfaceSDL2.cpp
@@ -405,7 +405,7 @@ int RmlUiSDL2SystemInterface::GetKeyModifiers()
 
 double RmlUiSDL2SystemInterface::GetElapsedTime()
 {
-	return SDL_GetTicks() / 1000;
+	return double(SDL_GetTicks()) / 1000.0;
 }
 
 bool RmlUiSDL2SystemInterface::LogMessage(Rml::Core::Log::Type type, const Rml::Core::String& message)
